+++ Gems/Atom/RPI/Code/Include/Atom/RPI.Public/Pass/Pass.h
@@ -12,6 +12,7 @@
 #include <Atom/RPI.Public/GpuQuery/GpuQuerySystemInterface.h>
 #include <Atom/RPI.Reflect/Image/Image.h>
 #include <Atom/RPI.Public/Image/AttachmentImage.h>
+#include <Atom/RPI.Public/Image/AttachmentImage.h>
 #include <Atom/RPI.Public/Pass/PassAttachment.h>
 #include <Atom/RPI.Public/Pass/PassDefines.h>
 #include <Atom/RPI.Public/Pass/PassSystemInterface.h>
@@ -59,6 +60,7 @@ namespace AZ
         struct PassRequest;
         struct PassValidationResults;
         class AttachmentReadback;
+        class ImageAttachmentCopy;
 
         using SortedPipelineViewTags = AZStd::set<PipelineViewTag, AZNameSortAscending>;
         using PassesByDrawList = AZStd::map<RHI::DrawListTag, const Pass*>;
@@ -93,6 +95,8 @@ namespace AZ
             : public AZStd::intrusive_base
         {
             AZ_RPI_PASS(Pass);
+
+            friend class ImageAttachmentPreviewPass;
 
         public:
             using ChildPassIndex = RHI::Handle<uint32_t, class ChildPass>;
@@ -369,6 +373,9 @@ namespace AZ
 
             void UpdateReadbackAttachment(FramePrepareParams params, bool beforeAddScopes);
 
+            // Setup ImageAttachmentCopy
+            void UpdateAttachmentCopy(FramePrepareParams params);
+
             // --- Protected Members ---
 
             const Name PassNameThis{"This"};
@@ -465,6 +472,9 @@ namespace AZ
             // For read back attachment
             AZStd::shared_ptr<AttachmentReadback> m_attachmentReadback;
             PassAttachmentReadbackOption m_readbackOption;
+
+            // For image attachment preview
+            AZStd::weak_ptr<ImageAttachmentCopy> m_attachmentCopy;
 
         private:
             // Return the Timestamp result of this pass
